#35 WIP: Add MeleeSkillAttackArbiter and skill checks

+++ src/Rhisis.World/Handlers/SkillHandler.cs
@@ -61,7 +61,7 @@ namespace Rhisis.World.Handlers
 
             SkillInfo skill = client.Player.SkillTree.GetSkillByIndex(packet.SkillIndex);
 
-            _skillSystem.UseSkill(client.Player, skill, (int)packet.TargetObjectId);
+            _skillSystem.UseSkill(client.Player, skill, packet.TargetObjectId);
         }
     }
 }
